% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{emoDataframeMaker}
\alias{emoDataframeMaker}
\title{EMO dataframe maker}
\usage{
emoDataframeMaker(text, sentimentType = "syuzhet", addColor = FALSE,
  nrc = FALSE)
}
\arguments{
\item{text}{Input text, typically a long-form document}

\item{sentimentType}{Type of sentiment analysis - an argument to syuzhet::get_sentiment, Default: 'syuzhet'}

\item{addColor}{Add a column with a colour code to indicate sentiment, Default: FALSE}

\item{nrc}{Add 8 columns with intensity scores for each of the emotions in the NRC emotion lexicon, Default: FALSE}
}
\value{
A dataframe where each row is a sentence with columns for the original text, sentiment score of the sentinence and cumulative sentiment score of the volume.
}
\description{
Given input text outputs a dataframe of sentiment scores.
}
\details{
Calculates sentence-by-sentence and cumulative sentiment scores for a given document.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[syuzhet]{get_sentences}},\code{\link[syuzhet]{get_sentiment}},\code{\link[syuzhet]{get_nrc_sentiment}}
}
