% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{emoPlotter}
\alias{emoPlotter}
\title{emoPlotter}
\usage{
emoPlotter(emoDF, showTrends = NULL, title = NULL, color = FALSE)
}
\arguments{
\item{emoDF}{One EMO dataframes}

\item{showTrends}{Optional input from slopeFinder, containing the best-fitting regression information, Default: NULL}

\item{title}{Optional string providing title for the plot. Default: NULL}

\item{color}{Show colours to indicate sentiment - requires EMO dataframe with colour data, Default: FALSE}
}
\value{
Returns a plot object
}
\description{
Takes one EMO dataframe and outputs a faceted plot of the emotional arcs
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
