% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nrcMultiPlotter}
\alias{nrcMultiPlotter}
\title{NRC plotter}
\usage{
nrcMultiPlotter(listOfEmos, titles = NULL)
}
\arguments{
\item{listOfEmos}{A list of EMO dataframes}

\item{titles}{Display the title of the document in the plot, Default: NULL}
}
\value{
A faceted plot of each document and its cumulative NRC scores
}
\description{
Given a list of EMO dataframes, plots the cumulative score for each emotion in Saif Mohammed's NRC emotion lexicon
}
\details{
Shows cumulative scores for anger, fear, anticipation, trust, surprise, sadness, joy, and disgust
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
