% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{slopeFinder}
\alias{slopeFinder}
\title{Slope finder}
\usage{
slopeFinder(emoDF)
}
\arguments{
\item{emoDF}{A document in the form of an EMO dataframe}
}
\value{
Outputs a tibble of one row reporting the best fitting story type.
}
\description{
Reports the fit of alternative emotional arcs to a document and the best fitting canonical story type.
}
\details{
The output tibble contains estimates of the linear slope and estimate, R2 for the linear and sinusoidal fit, the highest R2 achieved and the most appropriate category/story type
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[broom]{tidy}},\code{\link[broom]{glance}}
\code{\link[rlang]{syms}}
}
