% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{emoMultiPlotter}
\alias{emoMultiPlotter}
\title{EMO multiplotter}
\usage{
emoMultiPlotter(listOfEmos, titles = NULL, color = FALSE,
  showTrends = NULL)
}
\arguments{
\item{listOfEmos}{List of EMO dataframes}

\item{titles}{Show the document titles, Default: NULL}

\item{color}{Show colours to indicate sentiment - requires EMO dataframe with colour data, Default: FALSE}

\item{showTrends}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Takes a list of EMO  dataframes and outputs a faceted plot of the emotional arcs
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
